backend/requirements.txt
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
fastapi>=0.115.0
uvicorn[standard]
python-dotenv
pydantic>=2.10.0
scikit-learn
numpy
pandas
groq==0.5.0
python-multipart==0.0.9
httpx==0.27.0
scipy
joblib
